# Fix `slerp` and deprecate `linpol`