--- a/src/Quaternion.jl
+++ b/src/Quaternion.jl
@@ -312,7 +312,7 @@ end
     iszero(qa0) && throw(DomainError(qa0, "The input quaternion must be non-zero."))
     iszero(qb0) && throw(DomainError(qb0, "The input quaternion must be non-zero."))
     qa = qa0 / abs(qa0)
-    qb = qb0 / abs(qa0)
+    qb = qb0 / abs(qb0)
     coshalftheta = qa.s * qb.s + qa.v1 * qb.v1 + qa.v2 * qb.v2 + qa.v3 * qb.v3
 
     if coshalftheta < 0