This PR fixes #51.

## Deprecate `linpol`
As discussed in https://github.com/JuliaGeometry/Quaternions.jl/issues/51#issuecomment-1046887775, `slerp` seems a more common name.

## `slerp` now normalizes input quaternions
We had the following choices, and I choose the first one:
* Normalize all input quaternions, and the return value is always unit quaternion. (like `linpol`)
  * `linpol` is deprecated for its name, and the original `slerp` for non-unit quaternions seems buggy, so I thought it's better to have the same behavior as `linpol`.
* Throws `DomainError` for non-unit quaternion.
  * We will drop `norm` field with #75, and checking 
* Return `exp(log(q1)*(1-t) + log(q2)*t)` for any quaternions.
  * Slerp is shorthand for spherical linear interpolation, so it seems better to avoid non-unit quaternions output.

## The returned `Quaternion` has `norm` flag `true`.
The flag will be removed with #75, but seems better to fix the flag for now.

## TODO:
* [x] Add more tests
* [x] Check benchmark