@sethaxen 2022-03-05
> Return `exp(log(q1)*(1-t) + log(q2)*t)` for any quaternions.
> 
>     * Slerp is shorthand for spherical linear interpolation, so it seems better to avoid non-unit quaternions output.

I actually quite like this option. Although, I'll note that this expression is not quite correct. If `q1` and `q2` are on the same half-sphere, then `slerp(q1, q2, t)` is equivalent to `(q2 / q1)^t * q1`. If `q1 = a1 u1` and `q2 = a2 u2`, where `u1` and `u2` are unit quaternions and `a1>0` and `a2>0`, then `(q2 / q1)^t * q1 = a * slerp(u1, u2, t)`, where `log(a) = (1-t)*log(a1) + t * log(a2)`. i.e. this extension of `slerp` still spherically interpolates between the direction of the quaternions while it logarithmically interpolates between magnitudes, which seems sensible. I suspect it's more efficient still to normalize first and use `slerp` as implemented here, but the extension could be preserved by a single scalar multiplication at the end.