@hyrodium 2022-03-06
Thanks for the correction! The `exp(log(q1)*(1-t) + log(q2)*t)` was my mistake.
However,

>`(q2 / q1)^t * q1 = a * slerp(u1, u2, t)`, where `log(a) = (1-t)*log(a1) + t * log(a2)`

is also not quite correct with current `slerp` (and `linpol`) implementation.

```julia
julia> using Quaternions

julia> q1 = Quaternion(1.,0,0,0)
QuaternionF64(1.0, 0.0, 0.0, 0.0, false)

julia> q2 = Quaternion(-1/√2,-1/√2,0,0)
QuaternionF64(-0.7071067811865475, -0.7071067811865475, 0.0, 0.0, false)

julia> t = 1.0
1.0

julia> slerp(q1, q2, t)
QuaternionF64(0.7071067811865475, 0.7071067811865475, 0.0, 0.0, false)

julia> linpol(q1, q2, t)
QuaternionF64(0.7071067811865476, 0.7071067811865476, 0.0, 0.0, true)

julia> q1*(q2/q1)^t
QuaternionF64(-0.7071067811865474, -0.7071067811865475, 0.0, 0.0, false)
```
The current `slerp` and `linpol` equate antipodal points, but this does not seem to be sensible.