@hyrodium 2022-03-06
If we have the antipodal identification, then the magnitude interpolation might be useless, I guess. If someone needs the interpolation, one can just evaluate `q1*(q2/q1)^t` instead of `slerp`.