@sethaxen 2022-03-06
> The current `slerp` and `linpol` equate antipodal points, but I think this sometimes does not seem to be sensible.

When would this not be sensible? The unit quaternions doubly cover the rotations, so that `q` and `-q` correspond to the same rotation. While the unit quaternions form a 3-sphere, the rotations form a 3-hemisphere. The point of `slerp` is to smoothly interpolate between 2 rotations/orientations. Specifically, it computes the shortest path (geodesic) between the two _rotations_ and then traverses the specified distance on that path. One could compute a geodesic on the _unit quaternions_, but this wouldn't be slerp, and it wouldn't be all that useful.

> If we have the antipodal identification, then the magnitude interpolation might be useless, I guess.

I don't think so. The non-zero quaternions can be interpreted as the direct product of two manifolds: the (right-)quaternionic projective space (the hemisphere of unit quaternions corresponding to rotations) and the scale group (strictly positive reals). This is the manifold on which the `slerp` I'm proposing would be computing shortest paths.

> If someone needs the interpolation, one can just evaluate `q1*(q2/q1)^t` instead of `slerp`.

When I benchmarked this, I found it to be quite inefficient compared to `slerp`. Probably because it currently computes `log` and `exp` without any simplification.

I suppose I haven't argued for the benefits of doing it this way. Here are a few:
1. It's more general. I'd like to minimize how much rotation-specific code we have here, since quaternions are useful and interesting in their own right, not just as representations of rotations. If code has generalizations to the non-unit quaternions without much decrease in efficiency, it would be nice to implement the generalization.
2. It reduces sensitivity to normalization. A user doesn't need to worry about whether their quaternion is normalized before or after computing `slerp`. They can just normalize it whenever they need it to be normalized (which is probably right before computing a rotation matrix or rotating a pure quaternion). And, there might be a way to implement `slerp` this way that never requires computing the norm, and is therefore more efficient.
3. A person might want to use a quaternion to store both rotation and scale information. In this case, the generalization of `slerp` would just do the right thing.

I don't think this is _super_ important, and if you feel very strongly that `slerp` should only emit unit quaternions, then we can go that way.