@hyrodium 2022-03-08
>When would this not be sensible?

>One could compute a geodesic on the unit quaternions, but this wouldn't be slerp, and it wouldn't be all that useful.

The sphere S³ ≃ SU(2) ≃ (set of unit quaternions) is not just a double cover of SO(3) but also a universal cover. This means each point on S³ can be regarded as a path (or equivalence class of paths) in SO(3).

These videos might be helpful to understand the coverings:
https://vimeo.com/62228139
http://www.gregegan.net/APPLETS/21/21.html
https://youtu.be/8mhvnXWzlHw?t=146

Therefore, one quaternion can be interpreted as a path of rotation, so I think it would be useful to distinguish q and -q (with an optional argument like `shortest_path`).

>The non-zero quaternions can be interpreted as the direct product of two manifolds

>A person might want to use a quaternion to store both rotation and scale information.

I was thinking that unit quaternions is just a subset of ℍ. I now agree with the interpolation of magnitude (ℍ\{0} ≃ ℝ₊₀ × S³)! However, that will be a breaking change with `linpol`, so I'd like to keep the normalization in this PR.