@hyrodium 2022-03-21
To be clarify, I made two misunderstainding comments:

* `exp(log(q1)*(1-t) + log(q2)*t)`: this should be `q1*(q2/q1)^t`.
* I was thinking that the slerp should distinguish antipodal points `q` and `-q`.

I now think that it would be better to have a optional argument like `shortest_path` or `geodesic_path`. But this doesn't need to be in this PR.